aider: Set the height of the map container to a fixed value of 500px.
# Aider chat conversation:

USER: The map is still showing as disjointed tiles that are not shown correctly in a grid.

ASSISTANT: The issue with the disjointed map tiles might be due to the CSS styling of the map container. The map container needs to have a defined height for Leaflet to correctly position the map tiles.

To fix this, we can modify the CSS for the map container in `dashboard/frontend/sass/style.scss`:

1. Change the height of the map container to a fixed value.

Here is the *SEARCH/REPLACE* block:

dashboard/frontend/sass/style.scss
```scss
<<<<<<< SEARCH
#map {
    min-height: 500px;
    width: 100%;
    height: 0.9vh;
}
=======
#map {
    min-height: 500px;
    width: 100%;
    height: 500px;
}
>>>>>>> REPLACE
```

This change sets the height of the map container to 500px. You can adjust this value as needed to fit your layout.