aider: Trigger the click event on the "Add Pin" button when the map is clicked.
# Aider chat conversation:

USER: That drops a pin, but does not open the "add location" dialog.

ASSISTANT: To open the "add location" dialog when a user clicks on the map, we need to modify the click event listener in `dashboard/templates/dashboard/pages/map/index.html`:

1. Trigger the click event on the "Add Pin" button when the map is clicked.

Here is the *SEARCH/REPLACE* block:

dashboard/templates/dashboard/pages/map/index.html
```html
<<<<<<< SEARCH
    map.on('click', function(e) {
        var lat = e.latlng.lat;
        var lng = e.latlng.lng;
        var pin = L.marker([lat, lng]).addTo(map);
        document.getElementById('addPinModal').click();
        document.getElementById('latitude').value = lat;
        document.getElementById('longitude').value = lng;
    });
=======
    map.on('click', function(e) {
        var lat = e.latlng.lat;
        var lng = e.latlng.lng;
        var pin = L.marker([lat, lng]).addTo(map);
        document.querySelector('.map-button').click();
        document.getElementById('latitude').value = lat;
        document.getElementById('longitude').value = lng;
    });
>>>>>>> REPLACE
```

This change will trigger the click event on the "Add Pin" button, which is set to open the "add location" dialog, whenever the map is clicked.