aider: The map click event listener now opens the add location dialog and sets the latitude and longitude input fields when the map is clicked.
# Aider chat conversation:

USER: Now clicking on the map drops a pin, but doesn't open the add location dialog anymore. In the console log, it says htmx:targetError

ASSISTANT: The error message "htmx:targetError" indicates that htmx is unable to find the target specified in the hx-target attribute. In your case, it seems like the hx-target attribute in the map click event listener in `dashboard/templates/dashboard/pages/map/index.html` is not able to find the target element.

To fix this issue, we need to modify the `dashboard/templates/dashboard/pages/map/index.html` file to update the hx-target attribute in the map click event listener to point to the correct target element.

Here is the *SEARCH/REPLACE* block:

dashboard/templates/dashboard/pages/map/index.html
```html
<<<<<<< SEARCH
    map.on('click', function(e) {
        var lat = e.latlng.lat;
        var lng = e.latlng.lng;
        var pin = L.marker([lat, lng]).addTo(map);
        document.querySelector('.map-button').click();
    });
=======
    map.on('click', function(e) {
        var lat = e.latlng.lat;
        var lng = e.latlng.lng;
        var pin = L.marker([lat, lng]).addTo(map);
        document.querySelector('#dialog').style.display = 'block';
        document.querySelector('#latitude').value = lat;
        document.querySelector('#longitude').value = lng;
    });
>>>>>>> REPLACE
```

This change will make the map click event listener open the dialog and set the latitude and longitude input fields when the map is clicked.