aider: Fix the issue where the add location dialog doesn't appear when clicking on the map.
# Aider chat conversation:

USER: Now everything seems to work without errors, but when clicking on the map, the add location dialog doesn't appear.

ASSISTANT: To fix the issue where the add location dialog doesn't appear when clicking on the map, we need to modify the `dashboard/templates/dashboard/pages/map/index.html` file. The issue is likely due to the dialog element not being correctly targeted or displayed when the map is clicked.

Here is the *SEARCH/REPLACE* block:

dashboard/templates/dashboard/pages/map/index.html
```html
<<<<<<< SEARCH
    map.on('click', function(e) {
        var lat = e.latlng.lat;
        var lng = e.latlng.lng;
        var pin = L.marker([lat, lng]).addTo(map);
        var dialog = document.querySelector('#dialog');
        if (dialog) {
            dialog.style.display = 'block';
            document.querySelector('#latitude').value = lat;
            document.querySelector('#longitude').value = lng;
        }
    });
=======
    map.on('click', function(e) {
        var lat = e.latlng.lat;
        var lng = e.latlng.lng;
        var pin = L.marker([lat, lng]).addTo(map);
        var dialog = document.querySelector('#dialog');
        if (dialog) {
            dialog.style.display = 'block';
            document.querySelector('#latitude').value = lat;
            document.querySelector('#longitude').value = lng;
        } else {
            var addPinModal = document.querySelector('#addPinModal');
            if (addPinModal) {
                addPinModal.style.display = 'block';
                document.querySelector('#latitude').value = lat;
                document.querySelector('#longitude').value = lng;
            }
        }
    });
>>>>>>> REPLACE
```

This change will ensure that if the dialog element is not found, the script will try to find and display the addPinModal element instead. It will also set the latitude and longitude values in the addPinModal if it is found.