aider: The latitude and longitude input fields are set after the dialog has been fully loaded and displayed.
# Aider chat conversation:

USER: Now everything seems to work, but the very first time the map is clicked on and the dialog opens, the lat/lng isn't populated. The lat/lng is populated when the map is clicked a second time.

ASSISTANT: The issue you're experiencing is likely due to the fact that the map click event is triggering the "Add Pin" button click event, which in turn triggers the htmx request to open the dialog. However, the htmx request is asynchronous, meaning that the dialog may not be fully loaded and displayed by the time the map click event tries to set the latitude and longitude input fields.

To fix this issue, we can modify the `dashboard/templates/dashboard/pages/map/index.html` file to set the latitude and longitude input fields after the htmx request has completed and the dialog has been fully loaded and displayed. We can do this by listening for the `htmx:afterOnLoad` event, which is fired by htmx after an AJAX request has completed and the new content has been swapped into the DOM.

Here is the *SEARCH/REPLACE* block:

dashboard/templates/dashboard/pages/map/index.html
```html
<<<<<<< SEARCH
    map.on('click', function(e) {
        var lat = e.latlng.lat;
        var lng = e.latlng.lng;
        var pin = L.marker([lat, lng]).addTo(map);
        var addPinButton = document.querySelector('.map-button');
        if (addPinButton) {
            addPinButton.click();
            var latitudeInput = document.querySelector('#latitude');
            var longitudeInput = document.querySelector('#longitude');
            if (latitudeInput && longitudeInput) {
                latitudeInput.value = lat;
                longitudeInput.value = lng;
            }
        }
    });
=======
    map.on('click', function(e) {
        var lat = e.latlng.lat;
        var lng = e.latlng.lng;
        var pin = L.marker([lat, lng]).addTo(map);
        var addPinButton = document.querySelector('.map-button');
        if (addPinButton) {
            addPinButton.click();
            document.body.addEventListener('htmx:afterOnLoad', function() {
                var latitudeInput = document.querySelector('#latitude');
                var longitudeInput = document.querySelector('#longitude');
                if (latitudeInput && longitudeInput) {
                    latitudeInput.value = lat;
                    longitudeInput.value = lng;
                }
            }, { once: true });
        }
    });
>>>>>>> REPLACE
```

This change will ensure that the latitude and longitude input fields are set after the dialog has been fully loaded and displayed. The `{ once: true }` option ensures that the event listener is removed after it has been triggered once, so that the latitude and longitude input fields are not set every time an htmx request completes.