aider: The pin is now given an id when created and retrieved by its id when the dialog or screen is clicked.
# Aider chat conversation:

USER: This doesn't seem to work. I think that's probably because var pin is defined in index.html, but the javascript is defined in add_location.html. Make sure to pass the pin variable to the javascript in add_location.html somehow, for instance by setting an id on the pin, or a data attribute.

ASSISTANT: To fix the issue, we need to modify the `dashboard/templates/dashboard/pages/map/index.html` and `dashboard/templates/dashboard/pages/map/add_location.html` files.

Here are the steps:

1. In `dashboard/templates/dashboard/pages/map/index.html`, we need to add an id to the pin when it's created.
2. In `dashboard/templates/dashboard/pages/map/add_location.html`, we need to get the pin by its id when the dialog or screen is clicked.

Here are the *SEARCH/REPLACE* blocks:

dashboard/templates/dashboard/pages/map/index.html
```html
<<<<<<< SEARCH
    pin = L.marker([lat, lng]).addTo(map);
=======
    pin = L.marker([lat, lng], {id: 'pin'}).addTo(map);
>>>>>>> REPLACE
```

dashboard/templates/dashboard/pages/map/add_location.html
```html
<<<<<<< SEARCH
				  if (pin) {
				    map.removeLayer(pin);
				    pin = null;
				  }
=======
				  var pin = map.getLayerById('pin');
				  if (pin) {
				    map.removeLayer(pin);
				  }
>>>>>>> REPLACE
```

These changes add an id to the pin when it's created and get the pin by its id when the dialog or screen is clicked.